refactor(core): move zeus folder
