--- packages/core/src/graphql/buckets.ts
@@ -1,10 +1,0 @@
-export const UPSERT_BUCKET = `
-  mutation insertBucket($data: buckets_insert_input!) {
-    insertBucket(object: $data, on_conflict: {constraint: buckets_pkey}) {
-      id
-    }
-  }
-`;
-export interface InsertBucketParams {
-  id: string;
-}
