+++ packages/core/src/graphql/index.ts
@@ -1,5 +1,4 @@
-export * from './buckets';
-export * from './events';
-export * from './installations';
 export * from './types';
-export * from './users';
+export * from './zeus';
+export * from './zeus/typedDocumentNode';
+export * from './zeusScalars';
