+++ packages/core/src/graphql/zeus/typedDocumentNode.ts
@@ -1,23 +1,22 @@
 import { TypedDocumentNode } from '@graphql-typed-document-node/core';
 import gql from 'graphql-tag';
-
 import {
-  ExtractVariables,
+  ValueTypes,
   GenericOperation,
+  OperationOptions,
   GraphQLTypes,
   InputType,
-  OperationOptions,
   ScalarDefinition,
   ThunderGraphQLOptions,
-  ValueTypes,
   Zeus,
+  ExtractVariables,
 } from './';
 import { Ops } from './const';
 
 export const typedGql =
   <O extends keyof typeof Ops, SCLR extends ScalarDefinition, R extends keyof ValueTypes = GenericOperation<O>>(
     operation: O,
-    graphqlOptions?: ThunderGraphQLOptions<SCLR>
+    graphqlOptions?: ThunderGraphQLOptions<SCLR>,
   ) =>
   <Z extends ValueTypes[R]>(o: Z | ValueTypes[R], ops?: OperationOptions) => {
     const str = Zeus(operation, o, {
