+++ packages/core/src/index.ts
@@ -1,6 +1,3 @@
 export * from './functions';
 export * from './graphql';
 export * from './models';
-export * from './utils';
-export * from './zeus';
-export * from './zeus/typedDocumentNode';
